% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesReco-package.R
\docType{package}
\name{bayesReco-package}
\alias{bayesReco}
\alias{bayesReco-package}
\title{bayesReco: Probabilistic Reconciliation Via Conditioning}
\description{
Implementation of reconciliation methods for probabilistic hierarchical forecasts of time series.
}
\author{
\strong{Maintainer}: Dario Azzimonti \email{dario.azzimonti@gmail.com} (\href{https://orcid.org/0000-0001-5080-3061}{ORCID})

Authors:
\itemize{
  \item Nicolo' Rubattu \email{nicolo.rubattu@supsi.ch}
  \item Lorenzo Zambon \email{lorenzo.zambon@idsia.ch}
  \item Giorgio Corani \email{giorgio.corani@idsia.ch}
}

}
\keyword{internal}
