% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpart.R
\docType{data}
\name{carpart}
\alias{carpart}
\title{Carpart time series}
\format{
Univariate time series of class \link[stats]{ts}.
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008) Forecasting with exponential
smoothing: the state space approach, Springer
}
\usage{
carpart
}
\description{
A monthly time series from the \code{carparts} dataset, 51 observations, Jan 1998 - Mar 2002.
}
\keyword{datasets}
