% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carparts_example}
\alias{carparts_example}
\title{Example of a time series from carparts}
\format{
Univariate time series of class \link[stats]{ts}.
}
\source{
Godahewa, R., Bergmeir, C., Webb, G., Hyndman, R.J., & Montero-Manso, P. (2020).
Car Parts Dataset (without Missing Values) (Version 2)
\doi{10.5281/zenodo.4656021}
}
\usage{
carparts_example
}
\description{
A monthly time series from the \code{carparts} dataset, 51 observations, Jan 1998 - Mar 2002.
}
\references{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008).
Forecasting with exponential smoothing: the state space approach.
Springer Science & Business Media.

Godahewa, R., Bergmeir, C., Webb, G., Hyndman, R., & Montero-Manso, P. (2020).
Car Parts Dataset (without Missing Values) (Version 2)
\doi{10.5281/zenodo.4656021}
}
\keyword{datasets}
