% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc_MH.R
\name{.accept_prob}
\alias{.accept_prob}
\title{Compute acceptance probability}
\usage{
.accept_prob(b, b0, S, distr, params)
}
\arguments{
\item{b}{proposal state}

\item{b0}{current state}

\item{S}{aggregating matrix}

\item{distr}{list of strings specifying the distribution of each variable}

\item{params}{list of the parameters of the distributions}
}
\value{
the acceptance probability alpha
}
\description{
Compute acceptance probability
}
