% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{.check_hierarchical}
\alias{.check_hierarchical}
\title{Check if matrix is hierarchical}
\usage{
.check_hierarchical(A)
}
\arguments{
\item{A}{Aggregation matrix}
}
\value{
Logical value indicating if matrix is hierarchical
}
\description{
Returns TRUE if A is a hierarchy matrix.
}
\keyword{internal}
