% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{.get_Au}
\alias{.get_Au}
\title{Get aggregation matrix of upper sub-hierarchy}
\usage{
.get_Au(A, lowest_rows = NULL)
}
\arguments{
\item{A}{Aggregation matrix (must be hierarchical)}

\item{lowest_rows}{Optional. Integer vector of row indices corresponding to
the lowest upper level. If NULL, computed using \code{.lowest_lev(A)}.}
}
\value{
Aggregation matrix for the upper sub-hierarchy, or NULL if all
uppers are at the lowest level.
}
\description{
Constructs the aggregation matrix Au for the sub-hierarchy composed only
of the upper variables. This matrix relates the "upper uppers" to the
"lowest uppers" (the most disaggregated level of the upper hierarchy).
}
\keyword{internal}
