% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{.get_HG}
\alias{.get_HG}
\title{Extract and sort hierarchy rows}
\usage{
.get_HG(A, v, d, it)
}
\arguments{
\item{A}{Aggregation matrix}

\item{v}{Vector of values}

\item{d}{Distribution parameters}

\item{it}{Input type parameters}
}
\value{
List containing sorted hierarchy rows and remaining rows
}
\description{
Function that extracts the "best hierarchy rows" from A (see .get_hier_rows),
and sorts them in the correct order (i.e. bottom-up).
Also sorts accordingly the vectors v, d, it (e.g. of parameters).
}
\keyword{internal}
