% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{.lowest_lev}
\alias{.lowest_lev}
\title{Find rows corresponding to the lowest level}
\usage{
.lowest_lev(A)
}
\arguments{
\item{A}{Aggregation matrix (must be hierarchical)}
}
\value{
Integer vector of row indices corresponding to the lowest level
}
\description{
Identifies the rows of aggregation matrix A that correspond to the lowest
hierarchical level (i.e., the most disaggregated upper variables that
collectively sum to all bottom variables).
}
\keyword{internal}
