% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc_MH.R
\name{.proposal}
\alias{.proposal}
\title{Generate a proposal for MCMC step}
\usage{
.proposal(prev_prop, cov_mat_prop)
}
\arguments{
\item{prev_prop}{a list containing b, scale, acc_rate}

\item{cov_mat_prop}{the covariance matrix (ncol(cov_mat_prop)=length(b)) for the normal proposal}
}
\value{
a list containing the new b, scale, acc_rate
}
\description{
Generate a proposal for MCMC step
}
