% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infantMortality}
\alias{infantMortality}
\title{Infant Mortality grouped time series dataset}
\format{
List of time series of class \link[stats]{ts}.
}
\source{
hts package \href{https://cran.r-project.org/package=hts}{CRAN}
}
\usage{
infantMortality
}
\description{
A yearly grouped time series dataset, from 1901 to 2003, of infant mortality counts (deaths) in Australia;
disaggregated by state (see below), and sex (male and female).
}
\details{
States: New South Wales (NSW), Victoria (VIC), Queensland (QLD), South Australia (SA), Western Australia
(WA), Northern Territory (NT), Australian Capital Territory (ACT), and Tasmania (TAS).
}
\references{
Hyndman, R.J., Ahmed, R.A., Athanasopoulos, G., Shang, H.L. (2011).
Optimal combination forecasts for hierarchical time series.
Computational Statistics and Data Analysis, 55(9), 2579-2589.
}
\keyword{datasets}
