% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M3_example.R
\docType{data}
\name{M3_example}
\alias{M3_example}
\title{Example of a time series from the M3 forecasting competition}
\format{
List of time series of class \link[stats]{ts}.
}
\source{
\url{https://forecasters.org/resources/time-series-data/m3-competition/}
}
\usage{
M3_example
}
\description{
A monthly time series, from the M3 forecasting competition ("N1485").
}
\keyword{datasets}
