% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMF.R
\name{PMF.from_params}
\alias{PMF.from_params}
\title{Compute a PMF object from a parametric distribution}
\usage{
PMF.from_params(params, distr, Rtoll = .RTOLL)
}
\arguments{
\item{params}{a list of parameters:
\itemize{
\item lambda for the Poisson base forecast if \code{distr}='poisson', see \link[stats]{Poisson};
\item size and prob (or mu) for the negative binomial base forecast if \code{distr}='nbinom',
see \link[stats]{NegBinomial}.
}}

\item{distr}{a string that describes the distribution. Possible options "poisson" or "nbinom".}

\item{Rtoll}{Numerical tolerance for computing quantiles. Default .RTOLL=1e-9}
}
\value{
Samples drawn from the distribution specified by \code{pmf}.
}
\description{
Create a PMF object starting from the parametric form of a probability distribution.
}
\examples{
library(bayesRecon)

# Let's build the pmf of a negative binomial distribution with parameters p and size
prob <- 0.6
size <- 10 
pmf_nbinom <- PMF.from_params(params=list(prob=prob,size=size), distr="nbinom")

# Draw samples from the PMF object
set.seed(1)
samples <- PMF.sample(pmf=pmf_nbinom,N_samples = 1e4)

# Plot the histogram computed with the samples and the true value of the PMF
hist(samples,breaks=seq(-1,max(samples)),freq=FALSE)
points(seq(0,max(samples))-0.5,dnbinom(seq(0,max(samples)),size = size,prob = prob),pch=16)
points(seq(0,max(samples))-0.5,pmf_nbinom[1:(max(samples)+1)],pch=17)

}
\seealso{
\code{\link[=PMF.get_mean]{PMF.get_mean()}}, \code{\link[=PMF.get_var]{PMF.get_var()}}, \code{\link[=PMF.get_quantile]{PMF.get_quantile()}}, \code{\link[=PMF.summary]{PMF.summary()}}
}
