% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMF.R
\name{PMF.get_mean}
\alias{PMF.get_mean}
\title{Get the mean of the distribution from a PMF object}
\usage{
PMF.get_mean(pmf)
}
\arguments{
\item{pmf}{the PMF object.}
}
\value{
A numerical value for mean of the distribution.
}
\description{
Returns the mean from the PMF specified by \code{pmf}.
}
\examples{
library(bayesRecon)

# To do

}
\seealso{
\code{\link[=PMF.get_var]{PMF.get_var()}}, \code{\link[=PMF.get_quantile]{PMF.get_quantile()}}, \code{\link[=PMF.sample]{PMF.sample()}}
}
