% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMF.R
\name{PMF.get_quantile}
\alias{PMF.get_quantile}
\title{Get quantile from a PMF object}
\usage{
PMF.get_quantile(pmf, p)
}
\arguments{
\item{pmf}{the PMF object.}

\item{p}{the probability of the required quantile.}
}
\value{
A numeric value for the quantile.
}
\description{
Returns the \code{p} quantile from the PMF specified by \code{pmf}.
}
\examples{
library(bayesRecon)

# To do

}
\seealso{
\code{\link[=PMF.get_mean]{PMF.get_mean()}}, \code{\link[=PMF.get_var]{PMF.get_var()}}, \code{\link[=PMF.sample]{PMF.sample()}}
}
