% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMF.R
\name{PMF.get_var}
\alias{PMF.get_var}
\title{Get the variance of the distribution from a PMF object}
\usage{
PMF.get_var(pmf)
}
\arguments{
\item{pmf}{the PMF object.}
}
\value{
A numerical value for variance.
}
\description{
Returns the variance from the PMF specified by \code{pmf}.
}
\examples{
library(bayesRecon)

# To do

}
\seealso{
\code{\link[=PMF.get_mean]{PMF.get_mean()}}, \code{\link[=PMF.get_quantile]{PMF.get_quantile()}}, \code{\link[=PMF.sample]{PMF.sample()}}
}
