% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMF.R
\name{PMF.sample}
\alias{PMF.sample}
\title{Sample from the distribution given as a PMF object}
\usage{
PMF.sample(pmf, N_samples)
}
\arguments{
\item{pmf}{the PMF object.}

\item{N_samples}{number of samples.}
}
\value{
\code{N_samples} drawn from the distribution specified by \code{pmf}.
}
\description{
Samples (with replacement) from the probability distribution specified by \code{pmf}.
}
\examples{
library(bayesRecon)

# To do

}
\seealso{
\code{\link[=PMF.get_mean]{PMF.get_mean()}}, \code{\link[=PMF.get_var]{PMF.get_var()}}, \code{\link[=PMF.get_quantile]{PMF.get_quantile()}}
}
