% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc.R
\name{reconc_BUIS}
\alias{reconc_BUIS}
\title{Importance Sampling Reconciliation of Time Series Forecasts}
\usage{
reconc_BUIS(
  S,
  base_forecasts,
  in_type,
  distr,
  num_samples = 20000,
  seed = NULL
)
}
\arguments{
\item{S}{Summing matrix (n x n_bottom)}

\item{base_forecasts}{a list containing the base_forecasts, see details.}

\item{in_type}{A string with two possible values:
\itemize{
\item 'samples': base forecasts in the form of samples from the predictive distribution;
\item 'params' : base forecasts in the form of estimated parameters of a parametric predictive distribution.
}}

\item{distr}{A string describing the type of predictive distribution
\itemize{
\item 'continuous' or 'discrete' if \code{in_type}='samples'
\item 'gaussian', 'poisson' or 'nbinom' if \code{in_type}='params'
}}

\item{num_samples}{Number of samples}

\item{seed}{Seed for randomness reproducibility}
}
\value{
A list containing the reconciled forecasts. The list has the following named elements:
\itemize{
\item \code{bottom_reconciled_samples}: a matrix (n_bottom x \code{num_samples}) containing reconciled samples for the bottom time series
\item \code{upper_reconciled_samples}: a matrix (n_upper x \code{num_samples}) containing reconciled samples for the upper time series
\item \code{reconciled_samples}: a matrix (n x \code{num_samples}) containing the reconciled samples for all time series
}
}
\description{
Computation of reconciled forecasts for a hierarchy of time series with importance sampling.
}
\details{
The parameter \code{base_forecast} is a list containing n elements that depend on
the options \code{in_type} and \code{distr}.

If \code{in_type}='samples', each element of \code{base_forecast} is a vector containing samples from the base forecast distribution

If \code{in_type}='params', each element of \code{base_forecast} is a vector containing the estimated
\itemize{
\item mean and sd for the Gaussian base forecast, see \link[stats]{Normal}, if \code{distr}='gaussian';
\item lambda for the Poisson base forecast, see \link[stats]{Poisson}, if \code{distr}='poisson';
\item size and probability of success for the negative binomial base forecast, see \link[stats]{NegBinomial}, if \code{distr}='nbinom'.
}

The order of the \code{base_forecast} list is given by the order of the time series in the summing matrix.
}
\examples{
# Add poisson synthetic example

}
\references{
BUIS paper
}
\seealso{
\code{\link[=reconc_gaussian]{reconc_gaussian()}}
}
