% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc_probprog.R
\name{reconc_gaussian_sample}
\alias{reconc_gaussian_sample}
\title{Probabilistic reconciliation via STAN}
\usage{
reconc_gaussian_sample(
  base_forecasts,
  S,
  cores = NULL,
  chains = 4,
  iter = 2000,
  seed = NULL
)
}
\arguments{
\item{base_forecasts}{A list containing the base_forecasts, see details.}

\item{S}{Summing matrix (n x n_bottom)}

\item{cores}{Number of cores to use when executing the chains in parallel. If \code{NULL} use all available cores minus 1. See also \link[rstan]{sampling}.}

\item{chains}{A positive integer specifying the number of Markov chains. See also \link[rstan]{sampling}.}

\item{iter}{A positive integer specifying the number of samples for each chain. This number contains also the warm up iterations, by default \code{iter}/2. See also \link[rstan]{sampling}.}

\item{seed}{Seed for randomness reproducibility}
}
\value{
A list containing the reconciled forecasts. The list has the following named elements:
\itemize{
\item \code{reconciled_samples}: a matrix (n x (\code{iter} x \code{chains}/2 ) ) containing the reconciled samples for all time series
}
}
\description{
Probabilistic reconciliation via STAN
}
