% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc.R
\name{reconc_IS}
\alias{reconc_IS}
\title{This function bla bla bla...}
\usage{
reconc_IS(S, base_forecasts, in_type, distr, num_samples = 20000, seed = 42)
}
\arguments{
\item{S}{Summing matrix}

\item{base_forecasts}{base_forecasts}

\item{in_type}{A string 'samples'/'params'}

\item{distr}{A string 'continuous'/'discrete' or 'gaussian'/'poisson'/'nbinom'}

\item{num_samples}{number of samples}

\item{seed}{seed for randomness reproducibility}
}
\value{
Reconciled forecasts
}
\description{
This function bla bla bla...
}
