% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reconc_matrices}
\alias{reconc_matrices}
\title{A/S matrices building}
\usage{
reconc_matrices(aggf, bottom.f, bottom.H)
}
\arguments{
\item{aggf}{User-selected list of aggregates to consider.}

\item{bottom.f}{Integer seasonal period of the bottom time series.}

\item{bottom.H}{Bottom time series forecasting steps.}
}
\value{
A, S matrices
}
\description{
This function bla bla bla...
}
