% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc_MH.R
\name{reconc_MCMC}
\alias{reconc_MCMC}
\title{Probabilistic reconciliation with MCMC}
\usage{
reconc_MCMC(
  S,
  base_forecasts,
  distr,
  num_samples = 10000,
  tuning_int = 100,
  init_scale = 1,
  burn_in = 1000,
  seed = NULL
)
}
\arguments{
\item{S}{aggregating matrix}

\item{base_forecasts}{list of the parameters of the base forecast distributions}

\item{distr}{string or list of strings specifying the distribution of each variable}

\item{num_samples}{number of samples to draw using MCMC}

\item{tuning_int}{number of iterations between scale updates of the proposal}

\item{init_scale}{initial scale of the proposal}

\item{burn_in}{number of initial samples to be discarded}

\item{seed}{Seed for randomness reproducibility}
}
\value{
A list containing the reconciled forecasts. The list has the following named elements:
\itemize{
\item \code{bottom_reconciled_samples}: a matrix (n_bottom x \code{num_samples}) containing reconciled samples for the bottom time series
\item \code{upper_reconciled_samples}: a matrix (n_upper x \code{num_samples}) containing reconciled samples for the upper time series
\item \code{reconciled_samples}: a matrix (n x \code{num_samples}) containing the reconciled samples for all time series
}
}
\description{
Probabilistic reconciliation with MCMC
}
