% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc_TDcond.R
\name{reconc_TDcond}
\alias{reconc_TDcond}
\title{Probabilistic Reconciliation of forecasts via top-down conditioning}
\usage{
reconc_TDcond(
  S,
  fc_bottom,
  fc_upper,
  bottom_in_type = "pmf",
  distr = NULL,
  num_samples = 20000,
  return_type = "pmf",
  ...,
  suppress_warnings = FALSE,
  seed = NULL
)
}
\arguments{
\item{S}{Summing matrix (n x n_bottom).   return_pmf = TRUE, return_samples = FALSE, suppress_warnings = FALSE, seed = NULL}

\item{fc_bottom}{A list containing the bottom base forecasts, see details.}

\item{fc_upper}{A list containing the bottom base forecasts, see details.}

\item{bottom_in_type}{A string with three possible values:
\itemize{
\item 'pmf' if the bottom base forecasts are in the form of pmf, see details;
\item 'samples' if the bottom base forecasts are in the form of samples;
\item 'params'  if the bottom base forecasts are in the form of estimated parameters.
}}

\item{distr}{A string describing the type of bottom base forecasts ('gaussian', 'poisson' or 'nbinom').

This is only used if \code{bottom_in_type=='params'}.}

\item{num_samples}{Number of samples drawn from the reconciled distribution.}

\item{return_type}{The return type of the reconciled distributions. A string with three possible values:
\itemize{
\item 'pmf' returns a list containing reconciled pmf objects;
\item 'samples' returns a list containing reconciled samples;
\item 'all' returns a list with both pmf objects and samples.
}}

\item{...}{additional parameters to be passed to the smoothing functions for PMF objects.}

\item{suppress_warnings}{Logical. If \code{TRUE}, no warnings about samples
are triggered. If \code{FALSE}, warnings are generated. Default is \code{FALSE}. See Details.}

\item{seed}{Seed for reproducibility.}
}
\value{
A list containing the reconciled forecasts. The list has the following named elements:
\itemize{
\item \code{bottom_reconciled}: a list containing the pmf, the samples (matrix n_bottom x \code{num_samples}) or both, depending on the value of \code{return_type};
\item \code{upper_reconciled}: a list containing the pmf, the samples (matrix n_upper x \code{num_samples}) or both, depending on the value of \code{return_type}.
}
}
\description{
Uses the Top-down conditioning algorithm to draw samples from the reconciled
forecast distribution. Reconciliation is performed in two steps:
first, the upper base forecasts are reconciled via conditioning,
using only the hierarchical constraints between the upper variables; then,
the bottom distributions are updated via a probabilistic top-down procedure
}
\details{
The base (unreconciled) forecasts are passed with the parameters
\itemize{
\item \code{fc_bottom}: a list of length n_bottom where each element is either a pmf object (\code{bottom_in_type='pmf'}),
a vector of samples (\code{bottom_in_type='samples'}) or the parameters (\code{bottom_in_type='params'}) of the
parametric distribution specified in \code{distr}.
\item \code{fc_upper}: a list containing the parameters of the multivariate Gaussian distribution of the upper forecasts.
The list must contain only the named elements \code{mu} (vector of length n_upper) and \code{Sigma} (n_upper x n_upper matrix)
}

A PMF object is a numerical vector containing the probability mass function for a forecast.
The first element of a PMF vector for the variable X always contains the probability P(X=0) and
there is one element for each integer.
The last element of the PMF corresponds to the probability of the last value in the support of X.
See also \link{PMF.get_mean}, \link{PMF.get_var}, \link{PMF.sample}, \link{PMF.get_quantile} for functions that handle PMF objects.

A warnings is triggered if the intersection of the support for the reconciled uppers
and the support of the bottom-up distribution is too small. In this case only
few samples from the reconciled upper are kept. The warning reports the percentage
of samples kept.

Note that warnings are an indication that the base forecasts might have issues. Please check the base forecasts in case of warnings.
}
\examples{

library(bayesRecon)
# TO DO
}
\references{
Corani, G., Azzimonti, D., Augusto, J.P.S.C., Zaffalon, M. (2021). \emph{Probabilistic Reconciliation of Hierarchical Forecast via Bayes' Rule}. In: Hutter, F., Kersting, K., Lijffijt, J., Valera, I. (eds) Machine Learning and Knowledge Discovery in Databases. ECML PKDD 2020. Lecture Notes in Computer Science(), vol 12459. Springer, Cham. \doi{10.1007/978-3-030-67664-3_13}.

Zambon, L., Agosto, A., Giudici, P., Corani, G. (2023). \emph{Properties of the reconciled distributions for Gaussian and count forecasts}. \doi{10.48550/arXiv.2303.15135}.

Zambon, L., Azzimonti, D., Rubattu, N., Corani, G. (2024). \emph{Probabilistic reconciliation of mixed-type hierarchical time series} The 40th Conference on Uncertainty in Artificial Intelligence, accepted.
}
\seealso{
\code{\link[=reconc_BUIS]{reconc_BUIS()}}
}
