% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swiss_tourism}
\alias{swiss_tourism}
\title{Swiss tourism: monthly Swiss overnight stay data}
\format{
A list that contains:
\itemize{
\item \code{ts}: a multivariate time series of class \link[stats]{ts};
\item \code{n_bottom}: integer with the number of bottom time series (26);
\item \code{n_upper}: integer with the number of upper time series (1);
\item \code{agg_mat}: the aggregation matrix for the hierarchy. (1x26)
}
}
\source{
Federal Statistical Office of Switzerland (2025).
\emph{Hotel sector: arrivals and overnight stays of open establishments by year, month, canton and visitors' country of residence}.
\url{https://www.bfs.admin.ch/asset/en/px-x-1003020000_102}
}
\usage{
swiss_tourism
}
\description{
A dataset of 27 time series with monthly observations of the number of
overnight stays in Switzerland, from Jan 2005 to Jan 2025.
}
\details{
This is a hierarchy of time series that contains:
\itemize{
\item 26 bottom-level series corresponding to the 26 Swiss cantons
\item 1 top-level series corresponding to the Swiss total.
Each time series has 241 observations.
}
}
\references{
Carrara, C., Azzimonti, D., Corani, G., Zambon, L. (2025).
\emph{Modeling the uncertainty on the covariance matrix for probabilistic forecast reconciliation}.
arXiv:2506.19554
\doi{10.48550/arXiv.2506.19554}.
}
\keyword{datasets}
