% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temporal_aggregation}
\alias{temporal_aggregation}
\title{Time series temporal aggregation}
\usage{
temporal_aggregation(y, aggf = NULL)
}
\arguments{
\item{y}{Univariate time series of class ts.}

\item{aggf}{User-selected list of aggregates to consider.}
}
\value{
A list of aggregates time series.
}
\description{
This function bla bla bla...
}
