% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesReco-package.R
\docType{package}
\name{bayesReco-package}
\alias{bayesReco}
\alias{bayesReco-package}
\title{bayesReco: Probabilistic Reconciliation Via Conditioning}
\description{
Implementation of reconciliation methods for probabilistic hierarchical forecasts of time series.
}
\section{Main functions}{


The package implements reconciliation via conditioning for probabilistic forecasts of hierarchical time series. The main functions are
\itemize{
\item \link{reconc_IS}: a generic tool for reconciliation of any probabilistic time series forecast via importance sampling;
\item \link{reconc_gaussian}: implements analytic formulae for the reconciliation of Gaussian base forecasts.
}
}

\author{
\strong{Maintainer}: Dario Azzimonti \email{dario.azzimonti@gmail.com} (\href{https://orcid.org/0000-0001-5080-3061}{ORCID})

Authors:
\itemize{
  \item Nicolo' Rubattu \email{nicolo.rubattu@supsi.ch} (\href{https://orcid.org/0000-0002-2703-1005}{ORCID})
  \item Lorenzo Zambon \email{lorenzo.zambon@idsia.ch} (\href{https://orcid.org/0000-0002-8939-993X}{ORCID})
  \item Giorgio Corani \email{giorgio.corani@idsia.ch} (\href{https://orcid.org/0000-0002-1541-8384}{ORCID})
}

}
\keyword{internal}
