% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpart.R
\docType{data}
\name{carpart}
\alias{carpart}
\title{Carpart time series}
\format{
Univariate time series of class \link[stats]{ts}.
}
\source{
Godahewa, Rakshitha, Bergmeir, Christoph, Webb, Geoff, Hyndman, Rob, & Montero-Manso, Pablo. (2020). Car Parts Dataset (without Missing Values) (Version 2) \doi{10.5281/zenodo.4656021}
}
\usage{
carpart
}
\description{
A monthly time series from the \code{carparts} dataset, 51 observations, Jan 1998 - Mar 2002.
}
\references{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008) Forecasting with exponential
smoothing: the state space approach, Springer

Godahewa, Rakshitha, Bergmeir, Christoph, Webb, Geoff, Hyndman, Rob, & Montero-Manso, Pablo. (2020). Car Parts Dataset (without Missing Values) (Version 2) \doi{10.5281/zenodo.4656021}
}
\keyword{datasets}
