% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{get_reconc_matrices}
\alias{get_reconc_matrices}
\title{Build matrices for a hierarchy}
\usage{
get_reconc_matrices(aggf, bottom.f, bottom.h)
}
\arguments{
\item{aggf}{User-selected list of aggregates to consider.}

\item{bottom.f}{Integer seasonal period of the bottom time series.}

\item{bottom.h}{Bottom time series forecasting steps.}
}
\value{
A list containing the named elements
\itemize{
\item \code{A} the aggregation matrix;
\item \code{S} the summing matrix.
}
}
\description{
Creates the aggregation and summing matrices for a temporal hierarchy of time series from user-selected list of aggregations to consider.
}
\seealso{
\code{\link[=temporal_aggregation]{temporal_aggregation()}}
}
