% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc.R
\name{reconc_gaussian}
\alias{reconc_gaussian}
\title{Reconciliation in closed form for Gaussian base forecasts}
\usage{
reconc_gaussian(base_forecasts.mu, base_forecasts.Sigma, S)
}
\arguments{
\item{base_forecasts.mu}{base forecasts means vector}

\item{base_forecasts.Sigma}{base forecasts covariance matrix}

\item{S}{Summing matrix}
}
\value{
Reconciled mean and covariance matrix of the bottom and upper forecasts
}
\description{
This function bla bla bla...
}
