% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconc_MH.R
\name{reconc_MCMC}
\alias{reconc_MCMC}
\title{Probabilistic reconciliation with MCMC}
\usage{
reconc_MCMC(
  S,
  distr,
  params,
  N_samples = 10000,
  tuning_int = 100,
  init_scale = 1,
  burn_in = 1000,
  seed = NULL
)
}
\arguments{
\item{S}{aggregating matrix}

\item{distr}{list of strings specifying the distribution of each variable}

\item{params}{list of the parameters of the distributions}

\item{N_samples}{number of samples to draw using MCMC}

\item{tuning_int}{number of iterations between scale updates of the proposal}

\item{init_scale}{initial scale of the proposal}

\item{burn_in}{number of initial samples to be discarded}

\item{seed}{Seed for randomness reproducibility}
}
\value{
samples from the reconciled distribution
}
\description{
Probabilistic reconciliation with MCMC
}
