% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{temporal_aggregation}
\alias{temporal_aggregation}
\title{Non-overlapping temporal aggregation of a time series}
\usage{
temporal_aggregation(y, agg_levels = NULL)
}
\arguments{
\item{y}{univariate time series of class \link[stats]{ts}.}

\item{agg_levels}{user-selected list of aggregation levels.}
}
\value{
A list of \link[stats]{ts} objects each containing the aggregates time series in the order defined by \code{agg_levels}.
}
\description{
Creates a list of aggregated time series from a time series of class \link[stats]{ts}.
}
\details{
If \code{agg_levels=NULL} then \code{agg_levels} is automatically generated by taking all the factors of the time series frequency.
}
\seealso{
\link{get_reconc_matrices}
}
