% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{temporal_aggregation}
\alias{temporal_aggregation}
\title{Non-overlapping temporal aggregation of a time series}
\usage{
temporal_aggregation(y, aggf = NULL)
}
\arguments{
\item{y}{Univariate time series of class \link[stats]{ts}.}

\item{aggf}{User-selected list of aggregates to consider.}
}
\value{
A list of \link[stats]{ts} objects each containing the aggregates time series in the order defined by \code{aggf}.
}
\description{
Creates a list of aggregated time series from a time series of class \link[stats]{ts}.
}
\details{
If \code{aggf=NULL} then \code{aggf} is automatically generated by taking all the factors of the time series frequency.
}
\seealso{
\code{\link[=get_reconc_matrices]{get_reconc_matrices()}}
}
